Add modal JS file and Fix Linter Errors
