+++ src/css/style.css
@@ -1,0 +1,1 @@
+@import url("https://pro.fontawesome.com/releases/v5.10.0/css/all.css");
