+++ src/index.html
@@ -13,6 +13,33 @@
     ></script>
   </head>
   <body class="d-flex flex-column h-100">
+    <header>
+      <nav class="navbar navbar-expand-lg navbar-light bg-light">
+        <div class="container-fluid">
+          <a class="navbar-brand" href="#"
+            ><i class="fab fa-product-hunt d-inline-block"> </i> Pokescripts</a
+          >
+          <button
+            class="navbar-toggler"
+            type="button"
+            data-bs-toggle="collapse"
+            data-bs-target="#navbarNavAltMarkup"
+            aria-controls="navbarNavAltMarkup"
+            aria-expanded="false"
+            aria-label="Toggle navigation"
+          >
+            <span class="navbar-toggler-icon"></span>
+          </button>
+          <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
+            <div class="navbar-nav">
+              <a class="nav-link active" aria-current="page" href="#">Home</a>
+              <a class="nav-link" href="#">Types</a>
+              <a class="nav-link" href="#">Species</a>
+            </div>
+          </div>
+        </div>
+      </nav>
+    </header>
     <!-- Buttons trigger modal -->
     <div class="container p-5 d-flex justify-content-between">
       <button
